requirements.txt
streamlit
pillow
tensorflow
matplotlib
scikit-image